chore: Move ratatui display into its own file
